Develope

+++ src/main/java/Funssion/Inforum/common/utils/SecurityContextUtils.java
@@ -4,11 +4,11 @@ import org.springframework.security.core.context.SecurityContextHolder;
 
 public abstract class SecurityContextUtils {
 
-    public static Integer getUserId() {
+    public static Long getUserId() {
         String userId = SecurityContextHolder.getContext().getAuthentication().getName();
         if (userId == "anonymousUser") {
-            return 0;
+            return 0L;
         }
-        return Integer.valueOf(userId);
+        return Long.valueOf(userId);
     }
 }
